Simplify validation logic
